requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
python-dotenv==1.1.1
requests==2.32.5
resend==2.13.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
